+++ src/ewatercycle/base/model.py
@@ -327,6 +327,15 @@ class eWaterCycleModel(BaseModel, abc.ABC):
         return self._bmi.get_output_var_names()
 
     @property
+    def input_var_names(self) -> Iterable[str]:
+        """List of a model's input variables."""
+        return self._bmi.get_input_var_names()
+
+    def var_units(self, name: str) -> str:
+        """Return the given variable's units."""
+        return self._bmi.get_var_units(name)
+
+    @property
     def start_time_as_isostr(self) -> str:
         """Start time of the model.